moved old commandSystem remnants into more appropriate locations
