+++ src/main/talium/system/chatTrigger/cooldown/CooldownType.java
@@ -1,4 +1,4 @@
-package talium.system.commandSystem.repositories;
+package talium.system.chatTrigger.cooldown;
 
 /**
  * A Unit of a Cooldown, used in {@link talium.system.chatTrigger.cooldown.ChatCooldown} to build a value with a unit
