+++ src/main/talium/system/chatTrigger/triggerEngine/TriggerProvider.java
@@ -5,9 +5,9 @@ import org.springframework.stereotype.Service;
 import talium.system.chatTrigger.persistence.TriggerEntity;
 import talium.system.chatTrigger.cooldown.ChatCooldown;
 import talium.system.chatTrigger.persistence.TriggerService;
-import talium.system.commandSystem.repositories.ChatMessage;
-import talium.system.commandSystem.repositories.CooldownType;
-import talium.system.commandSystem.repositories.TwitchUserPermission;
+import talium.inputs.Twitch4J.ChatMessage;
+import talium.system.chatTrigger.cooldown.CooldownType;
+import talium.inputs.Twitch4J.TwitchUserPermission;
 
 import java.util.ArrayList;
 import java.util.HashMap;
